refactor(chat): isolate annotation highlight geometry
